src/embeddingworker/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
google-cloud-secret-manager==2.18.1 